# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stroopwafel Social API
  version: 1.2.0
  description: API-first social scheduling and publishing API for agents.
servers:
  - url: http://localhost:8080
security:
  - basicAuth: []
  - apiKeyHeader: []
paths:
...
  /api/v1/channels/bulk/disable:
    post:
      summary: Bulk disable channels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkChannelStatusRequest'
      responses:
        '200':
          description: Bulk disable result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkChannelStatusResponse'
        '400':
          $ref: '#/components/responses/APIError'
  /api/v1/channels/bulk/enable:
    post:
      summary: Bulk enable channels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkChannelStatusRequest'
      responses:
        '200':
          description: Bulk enable result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkChannelStatusResponse'
        '400':
          $ref: '#/components/responses/APIError'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    APIError:
      description: API error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  schemas:
...
    APIError:
      type: object
      required: [error, error_code]
      properties:
        error:
          type: string
        error_code:
          type: string
...
    Channel:
      type: object
      required: [id, type, display_name, status, created_at, updated_at]
      properties:
        id: { type: integer }
        type: { type: string }
        display_name: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_test_at: { type: string, format: date-time, nullable: true }
        last_error: { type: string, nullable: true }
        supports_media: { type: boolean }
        media_types:
          type: array
          items: { type: string }
        requires_media: { type: boolean }
...
    BulkChannelStatusRequest:
      type: object
      required: [channel_ids]
      properties:
        channel_ids:
          type: array
          items: { type: integer }
    BulkChannelStatusResponse:
      type: object
      required: [requested, succeeded, failed, updated]
      properties:
        requested: { type: integer }
        succeeded: { type: integer }
        failed: { type: integer }
        errors:
          type: array
          items: { type: string }
        updated:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
...